refactor: TastingInfo 위치 constants로 이동
