+++ src/entities/tasting-list/index.ts
@@ -1,2 +1,1 @@
 export { default as useTastingListQuery } from './model/hooks/useTastingListQuery';
-export type { TastingInfo } from './model/types';
