+++ src/entities/tasting-list/model/types.ts
@@ -1,14 +1,4 @@
-import { DrinkType } from '@/shared';
-
-export interface TastingInfo {
-  drinkId?: number;
-  drinkName?: string;
-  drinkTaste?: string; // 음료 맛
-  drinkFlavor?: string; // 음료 향
-  drinkColor?: string; // 음료 색
-  drinkImgUrl?: string;
-  drinkType?: DrinkType;
-}
+import { TastingInfo } from '@/shared';
 
 export interface TastingList {
   tastingList: TastingInfo[];
