+++ src/shared/constants/types.ts
@@ -10,3 +10,13 @@ export interface LocationInfo {
   address: string;
   detail: string;
 }
+
+export interface TastingInfo {
+  drinkId?: number;
+  drinkName?: string;
+  drinkTaste?: string; // 음료 맛
+  drinkFlavor?: string; // 음료 향
+  drinkColor?: string; // 음료 색
+  drinkImgUrl?: string;
+  drinkType?: DrinkType;
+}
