+++ src/widgets/post-meeting/model/types.ts
@@ -1,5 +1,4 @@
-import { DrinkType, LocationInfo } from '@/shared';
-import { TastingInfo } from '@/entities/tasting-list';
+import { DrinkType, LocationInfo, TastingInfo } from '@/shared';
 
 export default interface PostMeetingFormData {
   meetingTitle: string;
